refactor(git): share hosted review target admission across consumers
