+++ src/renderer/src/components/right-sidebar/source-control/review/hosted-review-push-target.ts
@@ -1,33 +1,8 @@
 import type { GitUpstreamStatus } from '../../../../../../shared/git-status-types'
-import type { GitPushTarget } from '../../../../../../shared/worktree/types'
 import type { HostedReviewState } from '../../../../../../shared/hosted-review'
 import { isPositiveHostedReviewNumber } from '../../../../../../shared/hosted-review'
 
-export function hasUsableHostedReviewPushTarget(args: {
-  pushTarget?: GitPushTarget
-  upstreamStatus?: GitUpstreamStatus
-  hasResolvableHostedReviewPushTargetLink?: boolean
-  branchName?: string
-}): boolean {
-  const identity = args.upstreamStatus?.upstreamIdentity
-  if (args.pushTarget) {
-    return (
-      args.upstreamStatus === undefined ||
-      (identity?.selector.kind === 'named-remote' &&
-        identity.selector.value === args.pushTarget.remoteName &&
-        identity.mergeRef === `refs/heads/${args.pushTarget.branchName}`)
-    )
-  }
-  if (args.hasResolvableHostedReviewPushTargetLink) {
-    // Older peers supply only a label; wait for authoritative target metadata.
-    return (
-      args.upstreamStatus?.hasUpstream === true &&
-      args.branchName !== undefined &&
-      identity?.mergeRef === `refs/heads/${args.branchName}`
-    )
-  }
-  return args.upstreamStatus?.hasConfiguredPushTarget === true
-}
+export { hasUsableHostedReviewPushTarget } from '../../../../../../shared/hosted-review-push-target-admission'
 
 export function hasResolvableHostedReviewPushTargetLink(args: {
   linkedGitHubPR?: number | null
